fix: events comments schema data
