+++ apps/shared-components/messages/messages-list/message/message.component.ts
@@ -1,7 +1,6 @@
 import { Component, ElementRef, EventEmitter, Injector, Input, OnInit, Output, ViewChild } from '@angular/core';
 import { FormBuilder, Validators } from '@angular/forms';
 import { InViewportDirective } from '@commudle/in-viewport';
-import { SeoService } from '@commudle/shared-services';
 import { faGrin } from '@fortawesome/free-regular-svg-icons';
 import { faCircle, faTrash } from '@fortawesome/free-solid-svg-icons';
 import { NoWhitespaceValidator } from 'apps/shared-helper-modules/custom-validators.validator';
@@ -51,7 +50,6 @@ export class MessageComponent implements OnInit {
     private fb: FormBuilder,
     // private userMessageReceiptHandlerService: UserMessageReceiptHandlerService,
     private injector: Injector,
-    private seoService: SeoService,
     private votesService: SVotesService,
   ) {
     this.replyForm = this.fb.group({
@@ -68,7 +66,6 @@ export class MessageComponent implements OnInit {
 
   ngOnInit(): void {
     this.getAllVotes();
-    this.seoSchema();
   }
 
   emitReply(): void {
@@ -100,51 +97,6 @@ export class MessageComponent implements OnInit {
     // if (messageId && visible) {
     //   this.userMessageReceiptHandlerService.addMessageReceipt(messageId, new Date());
     // }
-  }
-
-  seoSchema() {
-    this.seoService.setSchema({
-      '@context': 'https://schema.org',
-      '@type': 'DiscussionForumPosting',
-      headline: 'Comments',
-      text: this.removeHtmlTags(this.message.content),
-      author: {
-        '@type': 'Person',
-        name: this.message.user.name ? this.message.user.name : this.message.user.username,
-        url: `https://www.commudle.com/users/${this.message.user.username}`,
-      },
-      datePublished: this.message.created_at,
-      comment: this.getUserMessages(),
-    });
-  }
-
-  getUserMessages() {
-    const resultArray = [];
-
-    for (const userMessage of this.message.user_messages) {
-      if (userMessage) {
-        const transformedMessage = {
-          '@type': 'Comment',
-          text: this.removeHtmlTags(userMessage.content),
-          author: {
-            '@type': 'Person',
-            name: userMessage.user.name ? userMessage.user.name : userMessage.user.username,
-            url: `https://www.commudle.com/users/${userMessage.user.username}`,
-          },
-          datePublished: userMessage.created_at,
-        };
-
-        resultArray.push(transformedMessage);
-      }
-    }
-
-    return resultArray;
-  }
-
-  removeHtmlTags(content): string {
-    const parser = new DOMParser();
-    const doc = parser.parseFromString(content, 'text/html');
-    return doc.body.textContent || '';
   }
 
   onHoverEnter(id) {
